pyproject.toml
[project]
name = "matrix"
authors = [
    { name = "Brooke Chalmers", email = "breq@breq.dev" },
    { name = "Ava Silver", email = "ava@avasilver.dev" },
]
requires-python = ">=3.11"
version = "0.1.0"
dependencies = [
    "datadog==0.51.0",
    "flask==3.1.0",
    "gpiozero==2.0.1",
    "JSON-log-formatter==1.1",
    "pillow==11.1.0",
    "qrcode==8.0",
    "requests==2.32.3",
    "spotipy==2.19.0",
    "python-dotenv==1.0.1",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["matrix"]


[tool.ruff]
line-length = 120
